changes in the supplier section and also added the order details page and download section
