+++ src/SubscriptionPlan/subscription.module.css
@@ -30,7 +30,12 @@
     display: flex;
     flex-direction: column;
 }
-
+.subscriptionContainer{
+    display: flex
+    ;
+        flex-direction: column;
+        row-gap: 30px;
+}
 .logoSection {
     text-align: center;
     padding: 30px 0px 0px;
@@ -79,6 +84,7 @@
     padding: 30px 40px;
     display: flex;
     flex-direction: column;
+    justify-content: space-between;
     row-gap: 30px;
     border-radius: 50px;
     box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px !important;
