+++ src/SubscriptionPlan/thankyou.module.css
@@ -1,0 +1,72 @@
+.container{
+    background-color: #f9f9fa;
+    width: 100%;
+    height: 100vh;
+    overflow: scroll;
+    background-attachment: fixed;
+    scrollbar-width: none;
+    -ms-overflow-style: none;
+    display: flex;
+    justify-content: center;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif !important;
+}
+.container::-webkit-scrollbar{
+display: none;
+}
+.section{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    row-gap: 30px;
+}
+.logoSection{
+    text-align: center;
+    padding: 30px 0px 0px;
+}
+.logo{
+    width: 360px;
+    height: auto;
+}
+.card{
+    border: 2px solid transparent;
+    transition: border-color 0.3sease -in-out;
+    cursor: pointer;
+    width: 750px;
+    background: linear-gradient(to left, #f5f5f7, #e1f2ff);
+    padding: 40px;
+    display: flex;
+    flex-direction: column;
+    row-gap: 20px;
+    border-radius: 40px;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px !important;
+    align-items: center;
+    text-align: center;
+}
+.cardHeading{
+    font-size: 26px;
+    font-weight: 500;
+    color: #282F86;
+}
+.cardContent{
+    font-size: 18px;
+    font-weight: 500;
+}
+@media screen and (max-width: 767px) {
+    .card{
+       width: 100%;
+    }
+    .section{
+        width: 92%;
+    }
+}
+@media screen and (max-width: 575px) {
+    .card{
+      padding: 24px;
+    }
+    .logo{
+        width: 300px;
+    }
+  
+}
