+++ src/Supplier/components/Invoices/PendingInvoices/PendingInvoice.js
@@ -158,7 +158,7 @@ const PendingInvoice = ({ invoiceList, currentPage, totalInvoices, invoicesPerPa
                         hideFirstLastPages={true}
                     />
                     <div className='pagi-total'>
-                        <div>Total Items: {totalInvoices}</div>
+                        <div className='completed-pagi-total'>Total Items: {totalInvoices}</div>
                     </div>
                 </div>
             </div>
