+++ src/Supplier/components/Invoices/ProformaInvoices/ProformaList.js
@@ -147,7 +147,7 @@ const ProformaList = ({ invoiceList, currentPage, totalInvoices, invoicesPerPage
                         hideFirstLastPages={true}
                     />
                     <div className='pagi-total'>
-                        <div>Total Items: {totalInvoices}</div>
+                        <div className='completed-pagi-total'>Total Items: {totalInvoices}</div>
                     </div>
                 </div>
             </div>
