+++ src/Supplier/components/Orders/SupplierLogistics/logistics.css
@@ -1,0 +1,3 @@
+.react-date-picker__wrapper {
+    border: none !important;
+}
