+++ src/Supplier/components/Orders/SupplierLogistics/supplierlogistics.module.css
@@ -56,6 +56,7 @@
     flex-direction: column;
     flex-wrap: wrap;
     gap: 20px;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px !important;
 }
 
 .logisticesInputSection {
@@ -94,6 +95,7 @@
     display: flex;
     flex-direction: column;
     gap: 20px;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px !important;
 }
 
 .logisticesDimensionSection {
@@ -269,6 +271,7 @@
     display: flex;
     flex-direction: column;
     gap: 6px;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px !important;
 }
 
 .nameContainer {
